removida a logica das class Life, Petlife e prayer da main
