+++ Json/percent_key.json
@@ -1,5 +1,5 @@
 {
-    "life_percent": 40,
+    "life_percent": 35,
     "prayer_percent": 35,
-    "pet_life_percent": 40
+    "pet_life_percent": 45
 }
