+++ Json/teclas.json
@@ -1,7 +1,8 @@
 {
     "life_key": [
         "f5",
-        "f6"
+        "f6",
+        "2"
     ],
     "prayer_key": [
         "tab"
